userapi/requirements.txt

Django==3.0.6
django-filter==2.2.0
djangorestframework==3.11.0
